[DEV-98] refactor: 최신할일 목록과 list todo병합
