+++ src/components/dashboard/recently-todo/recently-todo-list.tsx
@@ -4,17 +4,17 @@ import { useEffect, useRef, useState } from "react";
 
 import { useSuspenseQuery } from "@tanstack/react-query";
 
+import ListTodo from "@/components/@common/list-todo";
 import { cn } from "@/lib/cn";
+import getQueryClient from "@/lib/get-query-client";
 import { getDashboardOptions } from "@/services/dashboard/query";
-
-import Goal from "./goal";
-import RecentlyTodoCheckbox from "./recently-todo-item";
-import TodoIcons from "./todo-icons";
+import { DashboardResponse } from "@/types/dashboard";
 
 export default function RecentlyTodoList() {
   const { data } = useSuspenseQuery(getDashboardOptions());
   const [isOverflowing, setIsOverflowing] = useState(false);
   const listRef = useRef<HTMLDivElement>(null);
+  const queryClient = getQueryClient();
 
   const { todos } = data;
 
@@ -24,6 +24,19 @@ export default function RecentlyTodoList() {
     }
   }, [todos]);
 
+  const toggleStatus = (id: number) => {
+    queryClient.setQueryData<DashboardResponse>(["dashboard"], (oldData) => {
+      if (!oldData) return oldData;
+      const clonedData = {
+        ...oldData,
+        todos: oldData.todos.map((todo) =>
+          todo.id === id ? { ...todo, isDone: !todo.isDone } : todo,
+        ),
+      };
+      return clonedData;
+    });
+  };
+
   return (
     <div
       ref={listRef}
@@ -31,16 +44,14 @@ export default function RecentlyTodoList() {
       role="region"
     >
       <ul className="flex flex-col gap-8 pr-8">
-        {todos.map((todo) => (
-          <li key={todo.id} className="text-sm">
-            <div className="flex items-center justify-between">
-              <RecentlyTodoCheckbox todo={todo} />
-              <TodoIcons todo={todo} />
-            </div>
-            {todo.goalInformation && (
-              <Goal goal={todo.goalInformation} isDone={todo.isDone} />
-            )}
-          </li>
+        {todos.map((todo, index) => (
+          <ListTodo
+            key={todo.id}
+            todo={todo}
+            index={index}
+            toggleStatus={toggleStatus}
+            showGoal
+          />
         ))}
       </ul>
       <div
